requirements-dev.txt
# Development dependencies
-r requirements.txt

# Code quality tools
black==23.7.0
isort==5.12.0
flake8==6.0.0
bandit==1.7.5
safety==2.3.5

# Testing tools
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
coverage==7.2.7

# Pre-commit hooks
pre-commit==3.3.3

# Django development tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Additional development tools
ipython==8.14.0
django-shell-plus==1.2.0
